poprawki na dzien dziecka
Z raportu Zuzy (+ naprawione, - do zrobienia, * jest opis)
1. Doadwanie nowej firmy
+ możliwe jest dodanie firmy o istniejącym już w bazie NIPie
+ usuwanie firmy odbywa się poprzez wciśnięcie jednego przycisku, bez możliwości anulowania
+ możliwe dodanie firmy bez nazwy (puste pole "Nazwa")
+ możliwe dodanie firmy o istniejącej już w bazie nazwie

2. Wystawianie zwolnienia:
+ możliwe jest pozostawienie pustego pola "Powód zwolnienia" (?)
+ możliwe jest wystawienie zwolnienia na dowolnie długi okres i  od daty dwolonie odległej w przód
+ nie można wystawić zwolnienia na jeden dzień (ta sama data)
- niemożliwa ponowne wystawienie zwolnienia, jeśli plik PDF z poprzednim jest wciąż otwarty
*(U mnie działa) jeśli firmę wyszukano po nazie, wciśnięcie "Zapisz" generuje dwa alerty: "Wystąpił błąd" oraz "Brak połączenia z bazą" mimo połączenia z internetem i poprawnych danych firmy (problem nie występuje przy wyszukiwaniu po NIPie tej samej firmy)

3. ???
- po wywaleniu aplikacji otwartym PDFem, wraz z ponownym jej uruchomieniem wyskakuje alert z błędem, ale nie uniemożliwia dalszej pracy programu (błąd prawdopodobnie zniknie wraz z usunięciem reszty błędów c:)

+++ aplikacja-przychodnia/aplikacja-przychodnia/Pages/AdminPage.xaml.cs
@@ -117,11 +117,19 @@ namespace aplikacja_przychodnia.Pages
         }
 
         private void ButtonFirm_Delete_Click(object sender, RoutedEventArgs e)
-        {
-            Firm firm = FirmView.SelectedItem as Firm;
-            FirmLocalDataBase.Remove(firm);
-            FirmLocalDataBase.Save();
-            RefreshFirmsView();
+        { 
+            
+            if (MessageBox.Show("Usunąć dane firmy?", "Potwierdzenie usuwania firmy", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.No)
+            {
+                //do no stuff
+            }
+            else
+            {
+                Firm firm = FirmView.SelectedItem as Firm;
+                FirmLocalDataBase.Remove(firm);
+                FirmLocalDataBase.Save();
+                RefreshFirmsView();
+            }
         }
     }
 }
